chore:move package global config file
