+++ backend/src/main/java/com/backend/rootbackend/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.backend.rootbackend.global;
+package com.backend.rootbackend.global.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
